Extracting survivor validation from model to a dedicated survivor validator
