+++ spec/examples.txt
@@ -1,37 +1,37 @@
 example_id                                                                | status | run_time        |
 ------------------------------------------------------------------------- | ------ | --------------- |
-./spec/unit/models/concerns/infection_report_state_machine_spec.rb[1:1:1] | passed | 0.01589 seconds |
-./spec/unit/models/concerns/infection_report_state_machine_spec.rb[1:1:2] | passed | 0.01509 seconds |
-./spec/unit/models/inventory_spec.rb[1:1:1:1:1]                           | passed | 0.03571 seconds |
-./spec/unit/models/inventory_spec.rb[1:1:1:2:1]                           | passed | 0.03947 seconds |
-./spec/unit/models/inventory_spec.rb[1:1:2:1:1]                           | passed | 0.06675 seconds |
-./spec/unit/models/inventory_spec.rb[1:1:2:2:1]                           | passed | 0.03964 seconds |
-./spec/unit/models/item_spec.rb[1:1:1:1]                                  | passed | 0.01113 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:1:1]                            | passed | 0.07012 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:1:2]                            | passed | 0.03357 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:1:3]                            | passed | 0.06434 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:1:4]                            | passed | 0.01403 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:1:5]                            | passed | 0.02597 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:2:1]                            | passed | 0.04677 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:2:2]                            | passed | 0.08429 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:3:1]                            | passed | 0.01909 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:3:2]                            | passed | 0.04857 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:4:1]                            | passed | 0.04065 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:4:2]                            | passed | 0.04017 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:4:3]                            | passed | 0.02238 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:4:4]                            | passed | 0.0295 seconds  |
-./spec/unit/models/survivor_spec.rb[1:1:1:4:5]                            | passed | 0.02322 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:5:1]                            | passed | 0.01 seconds    |
-./spec/unit/models/survivor_spec.rb[1:1:1:5:2]                            | passed | 0.03679 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:5:3]                            | passed | 0.01779 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:1:5:4]                            | passed | 0.0436 seconds  |
-./spec/unit/models/survivor_spec.rb[1:1:1:5:5]                            | passed | 0.00759 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:2:1]                              | passed | 0.05877 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:2:2]                              | passed | 0.05556 seconds |
-./spec/unit/models/survivor_spec.rb[1:1:2:3]                              | passed | 0.02814 seconds |
-./spec/unit/observers/survivor_observer_spec.rb[1:1]                      | passed | 0.09972 seconds |
-./spec/unit/use_cases/infection_report_spec.rb[1:1:1]                     | passed | 0.01514 seconds |
-./spec/unit/use_cases/infection_report_spec.rb[1:1:2]                     | passed | 0.87205 seconds |
-./spec/unit/use_cases/infection_report_spec.rb[1:1:3]                     | passed | 0.10097 seconds |
-./spec/unit/use_cases/infection_report_spec.rb[1:2:1]                     | passed | 0.12061 seconds |
-./spec/unit/use_cases/infection_report_spec.rb[1:2:2]                     | passed | 0.06293 seconds |
+./spec/unit/models/concerns/infection_report_state_machine_spec.rb[1:1:1] | passed | 0.00577 seconds |
+./spec/unit/models/concerns/infection_report_state_machine_spec.rb[1:1:2] | passed | 0.00802 seconds |
+./spec/unit/models/inventory_spec.rb[1:1:1:1:1]                           | passed | 0.03736 seconds |
+./spec/unit/models/inventory_spec.rb[1:1:1:2:1]                           | passed | 0.02679 seconds |
+./spec/unit/models/inventory_spec.rb[1:1:2:1:1]                           | passed | 0.07329 seconds |
+./spec/unit/models/inventory_spec.rb[1:1:2:2:1]                           | passed | 0.02068 seconds |
+./spec/unit/models/item_spec.rb[1:1:1:1]                                  | passed | 1.01 seconds    |
+./spec/unit/models/survivor_spec.rb[1:1:1:1:1]                            | passed | 0.04403 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:1:2]                            | passed | 0.03359 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:1:3]                            | passed | 0.02164 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:1:4]                            | passed | 0.01587 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:1:5]                            | passed | 0.02949 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:2:1]                            | passed | 0.03868 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:2:2]                            | passed | 0.01904 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:3:1]                            | failed | 0.08341 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:3:2]                            | passed | 0.02009 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:4:1]                            | passed | 0.065 seconds   |
+./spec/unit/models/survivor_spec.rb[1:1:1:4:2]                            | passed | 0.02011 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:4:3]                            | passed | 0.03651 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:4:4]                            | passed | 0.02585 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:4:5]                            | passed | 0.01136 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:5:1]                            | passed | 0.05333 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:5:2]                            | passed | 0.04065 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:5:3]                            | passed | 0.02426 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:5:4]                            | passed | 0.04262 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:1:5:5]                            | passed | 0.0341 seconds  |
+./spec/unit/models/survivor_spec.rb[1:1:2:1]                              | passed | 0.0575 seconds  |
+./spec/unit/models/survivor_spec.rb[1:1:2:2]                              | passed | 0.05183 seconds |
+./spec/unit/models/survivor_spec.rb[1:1:2:3]                              | passed | 0.1127 seconds  |
+./spec/unit/observers/survivor_observer_spec.rb[1:1]                      | passed | 0.07682 seconds |
+./spec/unit/use_cases/infection_report_spec.rb[1:1:1]                     | passed | 0.0481 seconds  |
+./spec/unit/use_cases/infection_report_spec.rb[1:1:2]                     | passed | 0.01593 seconds |
+./spec/unit/use_cases/infection_report_spec.rb[1:1:3]                     | passed | 0.05942 seconds |
+./spec/unit/use_cases/infection_report_spec.rb[1:2:1]                     | passed | 0.06777 seconds |
+./spec/unit/use_cases/infection_report_spec.rb[1:2:2]                     | passed | 0.07665 seconds |
